Added a build tag githubtests for running test involving github remote interactions with auth.
Need to provide the tag `githubtests` to run these
tests.
